system-info: Make the command-line driver relocatable
Look for shared libraries nearby: go up one level from ${bindir} to
${prefix}, then descend into ${libdir}. This will make it easier to run
in a possibly-broken environment, which seems valuable, since its
purpose is to detect broken environments.

The helpers are deliberately not linked like this: we must make sure
they are just as "fragile" as actual games, so that they can detect
broken library dependencies.

Signed-off-by: Simon McVittie <smcv@collabora.com>