Consolidate code to find and set up argv for helpers
This merges the code to find the various helpers. There are two broad
classes:

- Helpers from this source package, normally found in
  /usr/libexec/steam-runtime-tools-0
- Tools we have repurposed from other source packages (wflinfo and
  vulkaninfo), normally found via the $PATH

For the former class, we look in the SrtSystemInfo's overridden
helpers path (if set), or in $SRT_HELPERS_PATH (if set), or in
${libexecdir}/steam-runtime-tools-${API_MAJOR}. This is the same
behaviour we always had.

For the latter class, we now look in the SrtSystemInfo's overridden
helpers path (if set), or in $SRT_HELPERS_PATH (if set), or in
${libexecdir}/steam-runtime-tools-${API_MAJOR}, or in $PATH.
Previously, if the SrtSystemInfo had not overridden the helpers path,
we only looked in $PATH - but that makes it harder to make
steam-runtime-system-info do the right thing, using libraries and
helpers from the Steam Runtime but diagnosing the absence of the
Steam Runtime environment variables, when run from outside a Steam
Runtime environment.

As a side benefit, this gives us proper error handling in
_srt_get_helpers_path(), and reduces duplication.

Signed-off-by: Simon McVittie <smcv@collabora.com>