Add initial check for joysticks
With this commit we start to check if the plugged joysticks are with the
expected permissions.

In a canonical Steam installation we also expect to have R/W
permissions in `/dev/uinput` with the current logged in user.
So we also check the ACL of `/dev/uinput` to ensure that the current
user, and the mask, have the read and write permissions.

The old `can-write-uinput` check has been incorporated in the new
`joystick-uinput-issues`.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>