wrap: correctly handle bwrap mounting in case-insensitive filesystems
We often tell bubblewrap to mirror a certain directory from outside the
container to the inside.
When we are in a case insensitive filesystem this operation might fail
because bubblewrap reads `/proc/self/mountinfo` to check if the
directory was successfully mounted, but it might happen that the
directory is not listed in `mountinfo` because the kernel has put a
different case combination.

To workaround this issue we try to predict which case combination the
kernel is going to use (with fd and readlink(2)), and adjust the
bind-mount accordingly.

(Resolves: #47, steam-runtime#326)

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>