pressure-vessel: Use a shorter random string for the server socket
The AF_UNIX sockets that are used as rendezvous-point are limited to 108
bytes by the Linux kernel.

In order to be more flexible with the socket directory size, we avoid
using the 36 bytes long UUID as the server socket.
Instead we randomize a 12 bytes string that saves us 24 bytes, that we
can spare for the socket directory, but it is still long enough to
prevent a DOS attack.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>