flatpak-export: Try to handle absolute symlinks from /usr/
If an host system uses absolute symlinks in `/usr`, if we just
bind-mount the whole directory, these symlinks will be either broken or
will point to an unexpected location.

For example Gentoo used to have a symlink `/usr/lib` with `/usr/lib64`
as a target.

If we notice an absolute symlink that points to `/usr/*`, we can convert
it to a relative symlink to preserve its original target even inside the
container.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>