Silence misleading warnings about gameoverlayrenderer.so
When we need to run the main executable for a game, Steam usually
appends to the LD_PRELOAD the 32-bit and 64-bit version of
gameoverlayrenderer.so

However this leads to multiple warnings in the pressure-vessel log
because ld.so will complain that it was not able to preload the 32-bit
(or 64-bit) version of the library that was requested.

To avoid that we can special case the game overlay library and replace
it with a path that contains `${PLATFORM}`. In this way at runtime it
will always point to the library of the right ABI.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>