runtime: Delete old temporary copies in the background
After we have decided to delete an old temporary copy from var/, there
is actually no need to continue to hold locks, because the temporary
copy is never reused. If we put this in the background, we can continue
with other setup in parallel with deleting the temporary copy, saving
some time during setup.

Signed-off-by: Simon McVittie <smcv@collabora.com>